+++ src/forestutils.cpp
@@ -263,6 +263,7 @@ NumericVector treeFoliarBiomass(IntegerVector SP, NumericVector N, NumericVector
   NumericVector lb(ncoh);
   for(int i=0;i<ncoh;i++) {
     lb[i] = ((N[i]/10000)*afbt[i]*pow(std::min(200.0,dbh[i]), bfbt[i])*exp(cfbt[i]*std::min(ltba[i], 200.0)));
+    lb[i] = lb[i] * exp(-0.0001*N[i]);//Correct for high density packing
   }
   if(!NumericVector::is_na(gdd)) {
     NumericVector Sgdd = speciesNumericParameterWithImputation(SP, SpParams, "Sgdd");