+++ NEWS.md
@@ -10,7 +10,8 @@
 - Control flag to calculate fire hazard during 'spwb()' 'pwb()' and 'growth()' simulations
 - Integration of SurEau-Ecos v2.0 code as new transpiration mode ('Cochard')
 - Bug correction tissue moisture and water balance in granier's model
-- Allometries revised. Shrub foliar biomass/fuel limited by tree basal area.
+- Tree biomass allometries revised. 
+- Tree foliar biomass corrected for high density. Shrub foliar biomass/fuel limited by tree basal area.
 
 # Version 2.9.3
 - Bounded leaf area index